Major rewrite of most of the code
This commit includes a general rewrite and clean-up of most of the
code, although the general procedures do not change. Some of the
changes (not extensive) include:

- PMAP.py not only calls routines, but no actual code, apart from
the optparser is there.
- External required libraries have been reduced to astropy, numpy and
scipy. Some other libraries are used (colorama, sexutils) but the code
is no part of pdrLib (remember to acknowledge).
- All the tables created are astropy.table and are saved to disk using
either astropy.io.ascii (with the fixed_width format) or
astropy.io.votable.
- Deprecated libraries and code have been removed, although a lot
still remains in pdrLib.py.
- Now all the routines use the logger and the output is clearer.
- HI background can now be determined interactively using clfind2d, as
for FUV (this still require some testing).