Added functions to calculate pixel scale and pc per pixel.
Added two options to the configuration file to define the pixel
scale in arcsec. If those options are set to None, the scale is
calculated from the WCS of the FUV and HI images. In any case, the
values configOpts['FUV_Pix_Scale'] and configOpts['HI_Pix_Scale']
can be used to directly access those values.

From these, the options configOpts['FUV_PcPerPixel'] and
configOpts['HI_PcPerPixel'] are created with the number of pc a pixel
contains in the FUV and HI images.