Added compatibility with both astropy v0.2.4 and v0.3-dev
PMAP.py checks the version of astropy and adds a couple keywords
to the configOpts dictionary with it. Modified all the places where
there can be problems because of incompatibilities between both
versions. In some cases I have left only the 0.2.4 version as it
has not been properly deprecated in 0.3 (e.g., the Writer keyword for
astropy.io.ascii.write).