Tentative fix for the case when FUV & HI images have different sizes
This commit introduces a new option called matchFOV. When True
(default), the common FOV for the FUV and HI images is calculated at
the beginning of the pipeline. When CLFind is run, this common FOV is
used to cut the original image (the trimmed version is saved as
Trimmed_<original name>) so that regions are found only in that
region.

Additionally, a new rejection has been added to checkFUV.py. If a
region is too close to the border of the image, it is rejected.

This solution to the problem of mismatched image sizes is not ideal,
but it seems to work. A better solution should be found at some point.

CHECKS ARE REQUIRES TO ENSURE THAT THE RESULTS DO NOT CHANGE AFTER
THIS FIX.